refactor: move types to `types/api` directory
